refactor [DD-XXX] Restructured MVVM layers — extracted models, errors, and resources into dedicated folders
