+++ daily_done/Models/Stats/DayStat.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+struct DayStat: Identifiable {
+    let id: Date
+    let count: Int
+}
