+++ daily_done/Resources/Assets.xcassets/AppIcon.appiconset/Contents.json
