+++ daily_done/Resources/Assets.xcassets/Contents.json
