+++ daily_done/Resources/Assets.xcassets/backgroundSecondary.colorset/Contents.json
