+++ daily_done/Resources/Assets.xcassets/destructive.colorset/Contents.json
