+++ daily_done/Resources/Assets.xcassets/success.colorset/Contents.json
