+++ daily_done/Resources/Assets.xcassets/textSecondary.colorset/Contents.json
