+++ daily_done/Resources/Localizable.xcstrings
