+++ daily_done/Utilities/Errors/StatsError.swift
@@ -1,0 +1,12 @@
+import Foundation
+
+enum StatsError: LocalizedError {
+    case loadFailed(Error)
+
+    var errorDescription: String? {
+        switch self {
+        case .loadFailed:
+            return "Could not load statistics. Please try again."
+        }
+    }
+}
